requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.5.0
pillow>=9.0.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
tqdm>=4.65.0
timm>=0.9.0
transformers>=4.30.0
requests>=2.28.0
tensorboard>=2.12.0 